#3383, #3881 Create GCP provider with shared VPC and allow specifying subnets per region
Summary:
If users want to create shared VPC universes, they need to specify the host project of the shared
when creating provider (along with the VPC). When they create, the universe will exist on the host
project rather than the project YW is run on.

Users can also specify a subnet in `perRegionMetadata` when creating a provider. This is currently
API only and will be followed up by the UI change in a separate diff.

Test Plan:
Shared VPC:
- Set up YW on subproject. Create universe with shared VPC.

Multiple subnets in region:
Create a provider without bootstrapping and then run:
```bin/ybcloud.sh gcp network bootstrap --custom_payload '{"errorString":null,"providerUUID":"5ce99a14-e6b0-4a19-a230-3bc5edb6cabd","perRegionMetadata":{"us-west1":{"vpcId":null,"vpcCidr":null,"azToSubnetIds":null,"customImageId":null,"customSecurityGroupId":null}},"keyPairName":null,"sshPrivateKeyContent":null,"sshUser":null,"hostVpcId":null,"hostVpcRegion":null,"customHostCidrs":[],"destVpcId":"wesley-shared-vpc","subnetId":"wesley-shared-subnet"}'```

Reviewers: ram, bogdan

Reviewed By: bogdan

Subscribers: andrew, jenkins-bot, yugaware

Differential Revision: https://phabricator.dev.yugabyte.com/D8014

+++ managed/src/main/java/com/yugabyte/yw/commissioner/tasks/subtasks/cloud/CloudRegionSetup.java
@@ -108,13 +108,18 @@ public class CloudRegionSetup extends CloudTaskBase {
           throw new RuntimeException(errMsg);
         }
         List<String> zones = Json.fromJson(zoneInfo.get(regionCode).get("zones"), List.class);
-        List<String> subnetworks = Json.fromJson(zoneInfo.get(regionCode).get("subnetworks"), List.class);
-        if (subnetworks.size() != 1) {
-          region.delete();
-          throw new RuntimeException(
-              "Region Bootstrap failed. Invalid number of subnets for region " + regionCode);
+        String subnetId = taskParams().metadata.subnetId;
+        if (subnetId == null || subnetId.isEmpty()) {
+          List<String> subnetworks = Json.fromJson(
+            zoneInfo.get(regionCode).get("subnetworks"), List.class);
+          if (subnetworks.size() != 1) {
+            region.delete();
+            throw new RuntimeException(
+                "Region Bootstrap failed. Invalid number of subnets for region " + regionCode);
+          }
+          subnetId = subnetworks.get(0);
         }
-        String subnet = subnetworks.get(0);
+        final String subnet = subnetId;
         region.zones = new HashSet<>();
         zones.forEach(zone -> {
           region.zones.add(AvailabilityZone.create(region, zone, zone, subnet));
